Moved URL to update action, styled a bit
