--- src/main/resources/com/atlassian/bitbucket/jenkins/internal/jenkins/oauth/consumer/OAuthApplinkInfoAction/sidepanel.properties
@@ -1,1 +1,0 @@
-bitbucket.oauth.consumer.back=Back to Consumer List
