+++ src/main/resources/com/atlassian/bitbucket/jenkins/internal/jenkins/oauth/consumer/OAuthConsumerUpdateAction/applinkinfo.jelly
@@ -12,18 +12,38 @@
             <p>
                 ${%bitbucket.oauth.applinks.info}
             </p>
-
+            <br/>
             <j:set var="descriptor" value="${it.consumerDescriptor}"/>
             <j:scope>
-                <f:entry field="requestTokenUrl" title="${%bitbucket.oauth.applinks.request.title}">
+                <h3>Specific for the consumer key stuff</h3>
+
+                ${%bitbucket.oauth.applinks.key.title}
+                <f:entry field="consumerKey" >
+                    <f:readOnlyTextbox value="${it.consumerKey}"/>
+                </f:entry>
+
+                <h3>URLs you also need</h3>
+
+                <p>
+                ${%bitbucket.oauth.applinks.request.title}
+                <f:entry field="requestTokenUrl">
                     <f:readOnlyTextbox value="${it.requestTokenUrl}"/>
                 </f:entry>
-                <f:entry field="accessTokenUrl" title="${%bitbucket.oauth.applinks.access.title}">
+                </p>
+
+                <p>
+                ${%bitbucket.oauth.applinks.access.title}
+                <f:entry field="accessTokenUrl">
                     <f:readOnlyTextbox value="${it.accessTokenUrl}"/>
                 </f:entry>
-                <f:entry field="authorizeUrl" title="${%bitbucket.oauth.applinks.authorize.title}">
+                </p>
+
+                <p>
+                ${%bitbucket.oauth.applinks.authorize.title}
+                <f:entry field="authorizeUrl">
                     <f:readOnlyTextbox value="${it.authorizeUrl}"/>
                 </f:entry>
+                </p>
             </j:scope>
         </l:main-panel>
     </l:layout>
