+++ src/main/resources/com/atlassian/bitbucket/jenkins/internal/jenkins/oauth/consumer/OAuthConsumerUpdateAction/applinkinfo.properties
@@ -1,5 +1,6 @@
 bitbucket.oauth.applinks.title=Applink Setup URLs
-bitbucket.oauth.applinks.info=Copy paste the following fields into the Applink setup dialogue in Bitbucket:
+bitbucket.oauth.applinks.info=Setup information for consumer in applinks, blah blah blah
+bitbucket.oauth.applinks.key.title=Consumer Key
 bitbucket.oauth.applinks.request.title=Request Token URL
 bitbucket.oauth.applinks.access.title=Access Token URL
 bitbucket.oauth.applinks.authorize.title=Authorize URL
