+++ src/main/resources/com/atlassian/bitbucket/jenkins/internal/jenkins/oauth/consumer/OAuthConsumerUpdateAction/sidepanel.properties
@@ -1,2 +1,3 @@
 bitbucket.oauth.consumer.back=Back to Consumer List
+bitbucket.oauth.consumer.applink.info=Applink Setup Info
 bitbucket.oauth.consumer.delete=Delete
