+++ src/main/resources/com/atlassian/bitbucket/jenkins/internal/jenkins/oauth/consumer/OAuthGlobalConfiguration/index.jelly
@@ -17,6 +17,7 @@
                     <th>${%bitbucket.oauth.consumer.key}</th>
                     <th>${%bitbucket.oauth.consumer.name}</th>
                     <th>${%bitbucket.oauth.consumer.callback}</th>
+                    <th>${%bitbucket.oauth.applink.setup}</th>
                 </tr>
                 <j:forEach var="d" items="${it.consumers}">
                     <tr class="center">
@@ -32,6 +33,7 @@
                                 </j:otherwise>
                             </j:choose>
                         </td>
+                        <td><a href="${'consumer/'+d.consumerKey+'/applinkinfo'}">${%bitbucket.oauth.applink.setup}</a></td>
                     </tr>
                 </j:forEach>
             </table>
