+++ src/main/resources/com/atlassian/bitbucket/jenkins/internal/jenkins/oauth/consumer/OAuthGlobalConfiguration/index.properties
@@ -1,3 +1,4 @@
+bitbucket.oauth.applink.setup=Applink Setup Info
 bitbucket.oauth.consumer.key=Consumer Key
 bitbucket.oauth.consumer.name=Consumer Name
 bitbucket.oauth.consumer.secret=Consumer Secret
