+++ src/main/resources/com/atlassian/bitbucket/jenkins/internal/jenkins/oauth/consumer/OAuthGlobalConfiguration/sidepanel.jelly
@@ -4,7 +4,6 @@
     <l:side-panel>
         <l:tasks>
             <l:task icon="secure.gif" href="./tokens" title="${%bitbucket.oauth.token.revoke}" contextMenu="false"/>
-            <l:task icon="help.gif" href="./info" title="${%bitbucket.oauth.applink.setup}" contextMenu="false" />
         </l:tasks>
     </l:side-panel>
 </j:jelly>
