+++ src/main/resources/com/atlassian/bitbucket/jenkins/internal/jenkins/oauth/consumer/OAuthGlobalConfiguration/sidepanel.properties
@@ -1,2 +1,1 @@
 bitbucket.oauth.token.revoke=Revoke Access Tokens
-bitbucket.oauth.applink.setup=Applink Setup Info
